requirements.txt
diffusers
transformers
accelerate
torch
Pillow
tqdm
